requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.4
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0